refactor: return const.js file.
