+++ src/model/offers-model.js
@@ -1,5 +1,5 @@
 import Observable from '../framework/observable.js';
-import {UpdateType} from '../utils/common-utils.js';
+import {UpdateType} from '../const.js';
 
 
 export default class OffersModel extends Observable {
