+++ src/presenter/filter-presenter.js
@@ -1,5 +1,5 @@
 import {render, remove, replace} from '../framework/render.js';
-import {UpdateType} from '../utils/common-utils.js';
+import {UpdateType} from '../const.js';
 import FilterView from '../view/filter-view.js';
 
 export default class FilterPresenter {
