+++ src/presenter/new-event-presenter.js
@@ -1,5 +1,6 @@
 import {render, RenderPosition, remove} from '../framework/render.js';
-import {UserAction, UpdateType, isEscapeKey} from '../utils/common-utils.js';
+import {UserAction, UpdateType} from '../const.js';
+import {isEscapeKey} from '../utils/common-utils.js';
 import EventCreateView from '../view/event-create-view.js';
