+++ src/service/events-api-service.js
@@ -1,4 +1,4 @@
-import {HttpMethod, HttpRoute} from '../utils/common-utils.js';
+import {HttpMethod, HttpRoute} from '../const.js';
 import ApiService from '../framework/api-service.js';
